backend/app/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
python-dotenv
pydantic
pydantic[email]
requests